feat(jaeger): move to jaeger pkg
