+++ pkg/trace/module.go
@@ -1,14 +1,11 @@
 package trace
 
 import (
-	"github.com/alexfalkowski/go-service/pkg/trace/opentracing"
+	"github.com/alexfalkowski/go-service/pkg/trace/opentracing/jaeger"
 	"go.uber.org/fx"
 )
 
 var (
 	// JaegerOpenTracing for fx.
-	JaegerOpenTracing = fx.Invoke(opentracing.RegisterJaeger)
-
-	// Module for fx.
-	Module = fx.Options(JaegerOpenTracing)
+	JaegerOpenTracing = fx.Invoke(jaeger.Register)
 )
